backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
boto3==1.35.0
gunicorn==21.2.0
